Unify file explorer across project and conversation views
