+++ front/components/assistant/conversation/files_panel/FilesTab.tsx
@@ -1,11 +1,9 @@
-import type {
-  ConversationAttachmentRow,
-  FilePanelCategory,
-} from "@app/components/assistant/conversation/files_panel/types";
+import type { ConversationAttachmentRow } from "@app/components/assistant/conversation/files_panel/types";
+import type { FilePanelCategory } from "@app/components/file_explorer/types";
 import {
   CATEGORY_CONFIG,
   MIN_FILES_FOR_SEARCH,
-} from "@app/components/assistant/conversation/files_panel/utils";
+} from "@app/components/file_explorer/utils";
 import { useDebounce } from "@app/hooks/useDebounce";
 import { getFileTypeIcon } from "@app/lib/file_icon_utils";
 import { getFileProcessedUrl } from "@app/lib/swr/files";
