+++ front/hooks/useFileContent.ts
@@ -1,0 +1,31 @@
+import { clientFetch } from "@app/lib/egress/client";
+import { useSWRWithDefaults } from "@app/lib/swr/swr";
+import { normalizeError } from "@app/types/shared/utils/error_utils";
+
+export function useFileContent({
+  url,
+  disabled,
+}: {
+  url: string | null;
+  disabled?: boolean;
+}) {
+  const isDisabled = disabled || !url;
+
+  const { data, error } = useSWRWithDefaults<string | null, string>(
+    url,
+    async (u: string) => {
+      const response = await clientFetch(u);
+      if (!response.ok) {
+        throw new Error("Failed to fetch file content");
+      }
+      return response.text();
+    },
+    { disabled: isDisabled }
+  );
+
+  return {
+    fileContent: data ?? null,
+    isFileContentLoading: !error && data === undefined && !isDisabled,
+    fileContentError: error ? normalizeError(error) : null,
+  };
+}
